+++ docs/release_notes.rst
@@ -4,6 +4,6 @@ Release Notes
 v0.1
 ----
 
-* The initial release of counsyl-io.
+* The initial release of counsyl-storage-utils.
 * Provides a path factory that creates SwiftPath and OSPath objects.
 * Provides various utilities for file system procedures.